Creation Of Panels And Pause Menu
Created the Tool Panels which are a major mechanic within our game, I made it so you can open and close the panels as well as interact with any UI within the panels. I also created the pause menu that includes the exit button (the other buttons are under development) and it even stops time while active.