Score Implementation And Panel Functionality
I added the score manager to the game, so far all it does is go up by 10 when a button is press however we will later make it go up various values as the player completes tasks. I also began to add/complete functionality to the synthesizer, signal panel, and notepad, however some of the panel functionality (such as the synthesizer and signal panel) will have to be finished by some of my teammates as it corresponds with there department. other changes include disabling buttons while the pause menu is up, changed all panels to be set active instead of instantiated, and etc.